+++ deltafi-core/src/main/java/org/deltafi/core/configuration/K8sConfiguration.java
@@ -29,6 +29,7 @@ import org.deltafi.core.plugin.deployer.customization.PluginCustomizationRepo;
 import org.deltafi.core.plugin.deployer.customization.PluginCustomizationService;
 import org.deltafi.core.plugin.deployer.image.PluginImageRepositoryRepo;
 import org.deltafi.core.services.DeltaFiPropertiesService;
+import org.deltafi.core.services.EventService;
 import org.deltafi.core.snapshot.SystemSnapshotService;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnCloudPlatform;
 import org.springframework.boot.cloud.CloudPlatform;
@@ -59,8 +60,8 @@ public class K8sConfiguration {
     }
 
     @Bean
-    public DeployerService deployerService(DeltaFiPropertiesService deltaFiPropertiesService, KubernetesClient kubernetesClient, PluginImageRepositoryRepo imageRepositoryRepo, PluginCustomizationService pluginCustomizationService, PluginRegistryService pluginRegistryService, SystemSnapshotService systemSnapshotService) {
-        return new K8sDeployerService(deltaFiPropertiesService, kubernetesClient, imageRepositoryRepo, pluginCustomizationService, pluginRegistryService, systemSnapshotService);
+    public DeployerService deployerService(DeltaFiPropertiesService deltaFiPropertiesService, KubernetesClient kubernetesClient, PluginImageRepositoryRepo imageRepositoryRepo, PluginCustomizationService pluginCustomizationService, PluginRegistryService pluginRegistryService, SystemSnapshotService systemSnapshotService, EventService eventService) {
+        return new K8sDeployerService(deltaFiPropertiesService, kubernetesClient, imageRepositoryRepo, pluginCustomizationService, pluginRegistryService, systemSnapshotService, eventService);
     }
 
 }